PROJ5: TF-IDF! Is it working? I don't know!

+++ src/stlhashtableindex.h
@@ -13,12 +13,12 @@ class STLHashTableIndex : public Index
 {
 public:
     STLHashTableIndex(std::string);
-    virtual void add(const unsigned int, const std::string, const unsigned int);
+    virtual void add(const unsigned int, const std::string, const double);
     virtual void save();
     virtual void load();
     virtual void clear();
     virtual void find(std::string);
-    virtual std::map<unsigned int, unsigned int> findAll(std::string);
+    virtual std::map<unsigned int, double> findAll(std::string);
 private:
     std::mutex tableLock;
     std::unordered_map<std::string, entry> table;
